+++ src/Facades/Trustbird.php
@@ -13,6 +13,8 @@ use Trustbird\TrustbirdManager;
  * @method static \Trustbird\Assets\Managers\AssetsManager assets()
  * @method static \Trustbird\Teams\Managers\TeamsManager teams()
  * @method static \Trustbird\Risks\Managers\RisksManager risks()
+ * @method static \Trustbird\Policies\Managers\PoliciesManager policies()
+ * @method static \Trustbird\Suppliers\Managers\SuppliersManager suppliers()
  *
  * @see TrustbirdManager
  */