+++ src/TrustbirdServiceProvider.php
@@ -13,6 +13,10 @@ use Trustbird\Risks\Contracts\HasRisks;
 use Trustbird\Risks\Models\Risk;
 use Trustbird\Policies\Contracts\HasPolicies;
 use Trustbird\Policies\Models\Policy;
+use Trustbird\Suppliers\Contracts\HasSupplierRelations;
+use Trustbird\Suppliers\Contracts\HasSuppliers;
+use Trustbird\Suppliers\Models\Supplier;
+use Trustbird\Suppliers\Models\SupplierRelation;
 use Trustbird\Teams\Contracts\HasTeams;
 use Trustbird\Teams\Models\Team;
 use Trustbird\Workspaces\Contracts\HasWorkspaces;
@@ -57,6 +61,14 @@ final class TrustbirdServiceProvider extends ServiceProvider
             'policy' => [
                 'contract' => HasPolicies::class,
                 'default' => Policy::class,
+            ],
+            'supplier' => [
+                'contract' => HasSuppliers::class,
+                'default' => Supplier::class,
+            ],
+            'supplier_relation' => [
+                'contract' => HasSupplierRelations::class,
+                'default' => SupplierRelation::class,
             ],
         ];